requirements.txt
gspread==5.8.0
google-auth==2.22.0
requests==2.31.0
python-dotenv==1.0.0
ollama==0.1.6
numpy==1.26.4
